debugged build_json()

+++ api/gsn_api/gsndb/models.py
@@ -328,9 +328,9 @@ class Attendance(models.Model):
         on_delete = models.PROTECT,
     )
     entry_datetime = models.DateTimeField(default = timezone.now)
-    total_abs = models.IntegerField(null = True)
-    total_unexabs = models.IntegerField(null = True)
-    total_exabs = models.IntegerField(null = True)
+    total_abs = models.FloatField(null = True)
+    total_unexabs = models.FloatField(null = True)
+    total_exabs = models.FloatField(null = True)
     total_tardies = models.IntegerField(null = True)
     avg_daily_attendance = models.FloatField(null = True)
     term_final_value = models.BooleanField(default = False)
